Logging initialization moved to extra utils class
So it can be re-used by all runners.
